+++ lib/modules/network/network_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: network
-// Updated at: 2026-07-16 19:57:39
-// Commit: feat(network): add bi-directional event bus handler for network
+// Updated at: 2026-07-18 16:57:53
+// Commit: feat(network): implement shimmer loading skeleton in network
 
 class networkComponentState {
-  final String commitRef = '2901';
-  final String lastMessage = 'feat(network): add bi-directional event bus handler for network';
-  final DateTime timestamp = DateTime.parse('2026-07-16T19:57:39+05:00');
+  final String commitRef = '2929';
+  final String lastMessage = 'feat(network): implement shimmer loading skeleton in network';
+  final DateTime timestamp = DateTime.parse('2026-07-18T16:57:53+05:00');
 }